requirements.txt
scipy==1.8.0
opencv-python==4.5.5
h5py==3.6.0
PIL==9.0.1
pytorch==1.11.0
argparse==1.1
tensorly==0.7.0
